Split pseudofs.rs into several modules
